backend/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-rest-auth
Django
django-allauth
django-cors-headers
django-filter
django-rest-knox
django-solo
django-unfold
djangorestframework
djangorestframework_simplejwt
djoser
idna
oauthlib
pillow
pycparser
PyJWT
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
tzdata
urllib3

